[release] 중급 프로젝트 배포 2  (#117)
* [fix] 유저 랭킹 조금 더 변화가 잘 보일 수 있도록 수정 (#116)

---------

+++ src/main/java/com/codeit/project/deokhugam/batch/processor/RankProcessor.java
@@ -43,7 +43,7 @@ public class RankProcessor implements ItemProcessor<Object, Rank> {
       }
       case USER -> {
         UserStatDto user = (UserStatDto) dto;
-        score = user.reviewRankScore() * 0.5 + user.likeCount() * 0.2 + user.commentCount() * 0.3;
+        score = user.reviewRankScore().doubleValue() * 0.5 + user.likeCount() * 0.2 + user.commentCount() * 0.3;
         id = user.userId();
       }
       default -> throw new IllegalArgumentException("Unsupported RankTarget: " + rankTarget);
