+++ src/main/java/com/codeit/project/deokhugam/domain/user/dto/UserStatDto.java
@@ -1,8 +1,10 @@
 package com.codeit.project.deokhugam.domain.user.dto;
 
+import java.math.BigDecimal;
+
 public record UserStatDto(
         Long userId,
         Long likeCount,
         Long commentCount,
-        Double reviewRankScore
+        BigDecimal reviewRankScore
 ) {}
